move conversion utils to shared folder
