+++ shared/modules/conversion.utils.js
